+++ aiohttp_utils/negotiation.py
@@ -60,15 +60,18 @@ Example with httpie:
 Customizing negotiation
 =======================
 
-Renderers are just callables that receive a `request <aiohttp.web.Request>`
-data and return the rendered representation of that data.
+Renderers are just callables that receive a `request <aiohttp.web.Request>`,
+data to render, and the current ``handler`` function
+
+Renderers can return either the rendered representation of the data
+or a `Response <aiohttp.web.Response>`.
 
 Example:
 
 .. code-block:: python
 
     def render_text(request, data):
-        return data.encode('utf-8')
+        return data.encode(request.charset)
 
     # OR, if you need to parametrize your renderer, you can use a class
 
@@ -76,7 +79,7 @@ Example:
         def __init__(self, charset):
             self.charset = charset
 
-        def __call__(self, request, data):
+        def __call__(self, request, data, handler):
             return data.encode(self.charset)
 
     render_text_utf8 = TextRenderer('utf-8')
@@ -159,7 +162,7 @@ class JSONRenderer:
     """Callable object which renders to JSON."""
     json_module = pyjson
 
-    def __call__(self, request, data):
+    def __call__(self, request, data, handler):
         return self.json_module.dumps(data).encode('utf-8')
 
 #: Render data to JSON. Singleton `JSONRenderer`. This can be passed to the
@@ -196,14 +199,15 @@ def negotiation_middleware(
                 renderers=renderers,
                 force=force_negotiation
             )
+            request['selected_media_type'] = content_type
             response = yield from handler(request)
 
             if getattr(response, 'data', None):
                 # Render data with the selected renderer
                 if asyncio.iscoroutinefunction(renderer):
-                    render_result = yield from renderer(request, response.data)
+                    render_result = yield from renderer(request, response.data, handler)
                 else:
-                    render_result = renderer(request, response.data)
+                    render_result = renderer(request, response.data, handler)
             else:
                 render_result = response