+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 ==================
 
 - [negotiation,path_norm] *Backwards-incompatible*: Changed signatures of ``negotiation.setup`` and ``path_norm.setup`` to be more explicit. Both now take keyword-only arguments which are the same as the module's configuration keys, except lowercased, e.g. ``setup(app, append_slash=True, merge_slashes=True)``.
+- [negotiation] *Backwards-incompatible*: Renderers receive a third argument, ``handler`` (the next handler in the middleware chain).
 - [runner] Make ``run`` importable from top-level ``aiohttp_utils`` module.
 - [runner] Fix behavior when passing ``reload=False`` when ``app.debug=True``